+++ client/src/main/java/com/google/cloud/gcs/analyticscore/client/VersionHelper.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2025 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.gcs.analyticscore.client;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class VersionHelper {
+  private static final Logger LOG = LoggerFactory.getLogger(VersionHelper.class);
+  static final String VERSION = loadVersion();
+
+  private VersionHelper() {}
+
+  private static String loadVersion() {
+    String version = "unknown";
+    try (InputStream stream =
+        VersionHelper.class.getResourceAsStream(
+            "/META-INF/maven/com.google.cloud.gcs.analytics/client/pom.properties")) {
+      if (stream != null) {
+        Properties properties = new Properties();
+        properties.load(stream);
+        version = properties.getProperty("version", "unknown");
+      }
+    } catch (IOException e) {
+      LOG.warn("Failed to load client version", e);
+    }
+    return version;
+  }
+}
